Reject resource hint requests
With resource hints in asset discovery, there is no guarantee that
the request will result in content being responded.

Add `shouldRequestResolve` method to determine if a request should
continue or be aborted in pupetteer.

If an asset is a resource hint then there is a purpose within the
request header, if this exists and is of a specific value we can
detirmine the request shouldn't resolve

Assets in resource hints are not necessary for the page to render
correctly they are there to give the browser better performance in
production applications/websites for the purpose of percy this assets my
never be requested thus storing unused asssets.

Fixes #53