WIP: Add rollup to transpile `percy-agent.js`
This is to resolve the error where winston was being being included in
various webpack bundles. We change the `main` entry point of the
`@percy/agent` package so only the file that's ready for a browser is
bundled. Then we transpile that file fully with babel and rollup so
their webpack config isn't looking to compile it too.

This solves https://github.com/percy/percy-cypress/issues/58 fully.