build: Introduce rollup to bundle `percy-agent.js` for browsers (#124)
* Introduce rollup config

This is to resolve the error where winston was being being included in
various webpack bundles. We change the `main` entry point of the
`@percy/agent` package so only the file that's ready for a browser is
bundled. Then we transpile that file fully with babel and rollup so
their webpack config isn't looking to compile it too.

This solves https://github.com/percy/percy-cypress/issues/58 fully.

* Remove browser field, remove logger, test in other SDKs

Most of the errors stemmed from us using the logger (which is a node
logger) in our SDK utils, which is apart of our main export of the
package. This would cause bundlers to try and bundle that code and
fail because it couldn't find FS

* Remove changes to `sdk-utils.ts` file

That will be handled in a different PR and we're changing the way SDKs
import that file