Revert DOM clean up methods
Given this JSBin: https://jsbin.com/?html,console,output

When we serialize the CSSOM into the `<style>` tag that the in-memory
styles own and then _remove_ those styles, it triggers `deleteRule` on
those styles in the CSSOM. Which means we're removing the text we
injected AND all of the styles that live in memory. This breaks their
app.

With the input cleaning, when we remove the value or
checked attribute on the element, it's possible those elements lose
their values entirely. For example, when we remove the `checked`
attribute from checkboxes in the customers DOM, this will make the
checkbox _no longer checked_, which is changing the state of their
application. It would be better, in this world without an AST, to
leave the mutations in place. Removing the mutations causes more
issues than it tries to solve.