fix: Expose `PercyAgent` as a global always (#178)
Second time is the charm right? We're going to beta test this one but
it should do it. Setting `extend: true` in the rollup config allows
the bundler to extend the global namespace (window).

So the output now looks like `this.PercyAgent = ` (where `this` is the
window object)

I cut a beta release of this and tested it on the various SDKs (btw, we need a better process for this):

- https://github.com/percy/percy-cypress/compare/rd/agent-beta-4.4
- https://github.com/percy/percy-puppeteer/compare/rd/agent-beta-4.4
- https://github.com/percy/percy-nightwatch/compare/rd/agent-beta-4.4
- https://github.com/percy/percy-webdriverio/compare/rd/agent-beta-4.4
- https://github.com/percy/percy-nightmare/compare/rd/agent-beta-4.4
- https://github.com/percy/percy-protractor/compare/rd/agent-beta-4.4
- https://github.com/percy/percy-capybara/compare/rd/agent-beta-4.4

This one should do the trick 👍