fix: Disable HTTPS errors for asset discovery (#199)
* Add failing test snapshot

This will fail to snapshot with Percy (but the test suite will pass). We will
need to make sure we review any diffs that arise.

* Ignore HTTPS errors when doing asset discovery on snapshotservice

This is so the asset discovery service doesn't fail to capture assets for
snapshots that are done on invalid HTTPS sites. Without this the assets would be
missing from the snapshots (as seen in the previous commits percy build).