Apply form elements value _property_ as an _attribute always
There's no reason for us to check to see if the value attribute exists on a DOM
node already. This was probably a check to prevent us from modifying the
customers DOM when we were mutating in place. Now we're only modifying the
clone, so it's safe to _always_ set the attribute from the DOM elements
property.