ci: Remove loop over test cases for explicit test cases (#269)
* ci: Remove loop over test cases for explicit test cases

I'm not really a fan of loops in tests and this is a good example for why. The
two large resource snapshots in that loop took 2055ms & 2186ms, which takes
almost all of the timeout. So then the following 3-4 files in that loop have
like 400ms to work with to do everything. Sometimes they can complete, other
times it fails because it hit the timeout.

This commit makes all of the tests their own asserts so they get a full 5000ms
to work with individually.

* Fix duplicate snapshot names