fix: Static service wait for network idle (#270)
I've been seeing a lot of errors like:

```
Error: Execution context was destroyed, most likely because of a navigation.
```

This happens because we try to start altering the page (by injecting our JS)
before everything is settled down. This change won't promise the page has
actually loaded (that's very hard to determine) but it will wait for most
network requests to resolve before trying do anything.