fix: Stop and finalize build on exec termination (#332)
* Fix incorrect run check

This property is a function and not a getter. Without invoking it as a function,
it's value is always truthy.

* Stop the agent service when the exec process is terminated

* Do not handle stopping exec multiple times

When the process was terminated, all of the signal handlers were triggered
causing multiple logs to be outputted. Skipping stopping the agent once another
event has stopped it causes subsequent events to skip straight to exiting, and
never finalizes builds.

Exiting is tracked so that if it is already being handled, there is no need to
handle exiting again.

* Correct test for PERCY_ENABLED

The previous test tested that PERCY_ENABLED _did not_ work. That is, it tested
that a warning was still thrown even though it was disabled.

That test has been corrected to assert that _nothing_ is thrown when disabled.

* Fix comment typo