test: Make ignore snapshot test more explicit (#333)
* test: Make ignore snapshot test more explicit

We randonly get test fails from the static snapshot commands `_buildPageUrls`
unit test. An example of the failure can be seen here:
https://circleci.com/gh/percy/percy-agent/2767

This doesn't exactly fix the issue, but it will make the error clear when it
occurs the next time. Right now we just now the array doesn't deeply equal, but
we're not sure _which_ item it is.

* Add assert for length of `pages` array