Pass snapshot logger on to service handlers
A multitude of services are involved when taking a snapshot. The agent service
creates the snapshot logger when a snapshot is taken, it then passes it onto the
snapshot service, which passes it to the asset discovery service, which passes
it to the resource service, which then passes it to the response service.