feat: Automatically capture Winston logs per-snapshot (#343)
* Update winston to 3.x

The biggest change is how formatting is applied. Winston 3 does not have default
formatting and format-specific options like showLevel and label have been moved
into a more robust format syntax.

* Capture and upload per-snapshot logs

* Pass snapshot logger on to service handlers

A multitude of services are involved when taking a snapshot. The agent service
creates the snapshot logger when a snapshot is taken, it then passes it onto the
snapshot service, which passes it to the asset discovery service, which passes
it to the resource service, which then passes it to the response service.

* Tell winston to send error logs to stderr

In winston 3.x this is no longer the default behavior