fix: start detached and stop cleanup (#403)
* fix: :bug: healthcheck default port when provided port is undefined

Also use the HEALTHCHECK_PATH constant in this helper

* fix: :bug: do not pass "undefined" into detached start process

* fix: :bug: cleanup PID file on stop rather than kill the process

The process has handlers to finalize on SIGHUP, but the stop request already
finalizes the build before sending SIGHUP. Sometimes this caused duplicate,
empty, builds due to the finalize method creating a build when the current build
has already been finalized.

* :white_check_mark: Add separate test for no flags passed to start

* fix: :bug: Use oclif exit method

Testing with `process.exit` causes the test process to exit

* fix: :bug: await on subprocess and exit gracefully

This command was hard to test due to the subprocess being spawn async. In
practice the parent process remains open as well, but for our tests this method
returned before the command was finished.

* :white_check_mark: Update helpers to not throw oclif exit errors

* :white_check_mark: Fix snapshot command tests

These tests previously caused the entire suite to exit early. Once that was
fixed, the two tests written with the oclif helper failed.

One was updated to not need the helper, and the other test was removed since the
snapshot command now has a default directory argument.

* :white_check_mark: Fix agent tests and remove global mutation

This agent test caused many other tests to fail since it mutated the
DEFAULT_CONFIGURATION object.

* :white_check_mark: Unskip exec tests

With recent bug fixes, these tests can be unskipped and actually run.

The 'echo' test was replaced with 'sleep' to avoid printing to the inherited
stdio of the test process.

An unfinished test was fleshed out, and the other unfinished test removed since
exec does not support that behavior.

* :wrench: Remove silent junit reporter

When the tests exited early or threw warnings and logs, it went unnoticed in CI
since junit reports to a file. Circle only shows tests when they fail, so passed
tests or missed tests were never shown.