fix: Call `toString` on error
Issue #478 attempted to break the error apart to perserve the stacktrace, but it
looks like it's failing to produce the error message at all (depedning on the
test framework/logging setup).

We're going to call `.toString` on the error to print the message. If the user
wants the stack trace, they will need to open the browser and place a breakpoint
on the error log.