fix(🐛): upgrade puppeteer to try fixing page crashes (#539)
* fix(🐛): upgrade puppeteer to try fixing page crashes

It isn't guaranteed that this will prevent page crashes, but it's worth a shot. If this doesn't fix
the issue, then we might need to consider downgrading back to Puppeteer 2.x

* fix(📌): pin puppeteer to 5.0.0

The revisions of Chromium included with puppeteer since 5.0.0 seem to have issues with woff2
requests. Need to investigate further before upgrading to the latest version to see if it's a
breaking change or a bug.